+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # stoker
 
+## 0.4.264
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/web-app@0.5.222
+
 ## 0.4.263
 
 ### Patch Changes